chore(tests): create response assertion trait
